+++ src/admin/listeners.js
@@ -31,6 +31,7 @@ import {
   fetchLogChannel,
   gifParser,
   hasOctagonalSign,
+  isProduction,
   parseUnixTimestamp,
   setupEmbed,
 } from "../helpers/index.js";
@@ -67,9 +68,9 @@ export const onChannelUpdate = async (oldChannel, newChannel) => {
   const perm = chnUp.permissionOverwrites;
 
   //basic operations
-  if (process.env.DEBUG === "no" && isTestServer(newChannel)) return; //if in prod && modif in test server
-  const logChannel = await fetchLogChannel(newChannel); //get logChannelId
-  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
+  if (isProduction && isTestServer(newChannel)) return; //if in prod && modif in test server
+  const logChannel = await fetchLogChannel(newChannel.guild); //get logChannelId
+  if (isProduction && isTestServer(logChannel)) return; //if in prod && modif in test server
 
   //get client
   const client = newChannel.client;
@@ -314,9 +315,9 @@ export const onThreadCreate = async (thread, newly) => {
   if (thread) {
     //sometimes thread is null
     if (thread.joinable && !thread.joined) await thread.join(); //join thread created
-    if (process.env.DEBUG === "no" && isTestServer(thread)) return; //if in prod && modif in test server
+    if (isProduction && isTestServer(thread)) return; //if in prod && modif in test server
 
-    const logChannel = await fetchLogChannel(thread); //get logChannel
+    const logChannel = await fetchLogChannel(thread.guild); //get logChannel
     const perso = PERSONALITY.getAdmin().threadCreate;
     const log = await fetchAuditLog(
       thread.guild,
@@ -380,9 +381,9 @@ export const onRoleUpdate = async (oldRole, newRole) => {
   const roleUp = personality.roleUpdate;
   const auditLog = personality.auditLog;
 
-  if (process.env.DEBUG === "no" && isTestServer(newRole)) return; //if in prod && modif in test server
-  const logChannel = await fetchLogChannel(newRole); //get logChannelId
-  if (process.env.DEBUG === "no" && isTestServer(logChannel)) return; //if in prod && modif in test server
+  if (isProduction && isTestServer(newRole)) return; //if in prod && modif in test server
+  const logChannel = await fetchLogChannel(newRole.guild); //get logChannelId
+  if (isProduction && isTestServer(logChannel)) return; //if in prod && modif in test server
 
   //get client
   const client = newRole.client;
@@ -484,8 +485,8 @@ export const onMessageDelete = async (message) => {
   const messageDel = personality.messageDelete;
   const auditLog = personality.auditLog;
 
-  if (process.env.DEBUG === "no" && isTestServer(message)) return; //if in prod && modif in test server
-  const logChannel = await fetchLogChannel(message, "thread"); //get logChannel
+  if (isProduction && isTestServer(message)) return; //if in prod && modif in test server
+  const logChannel = await fetchLogChannel(message.guild, "thread"); //get logChannel
 
   const uDate = new Date(message.createdAt); //set date as Date object
   if (currentServer.name === "prod") uDate.setHours(uDate.getHours() + 1); //add 1h to date
@@ -670,8 +671,8 @@ export const onMessageUpdate = async (oldMessage, newMessage) => {
   const messageU = personality.messageUpdate;
   const auditLog = personality.auditLog;
 
-  if (process.env.DEBUG === "no" && isTestServer(newMessage)) return; //if in prod && modif in test server
-  const logChannel = await fetchLogChannel(nMessage, "thread"); //get logChannel
+  if (isProduction && isTestServer(newMessage)) return; //if in prod && modif in test server
+  const logChannel = await fetchLogChannel(nMessage.guild, "thread"); //get logChannel
 
   const color = Colors.DarkGreen;
   const embed = setupEmbed(color, messageU, nMessage.author, "tag"); //setup embed
@@ -827,10 +828,10 @@ export const onGuildBanAdd = async (userBan) => {
   const personality = PERSONALITY.getAdmin(); //get personality
   const perso = personality.guildBan;
 
-  if (process.env.DEBUG === "no" && isTestServer(userBan)) return; //if in prod && modif in test server
+  if (isProduction && isTestServer(userBan)) return; //if in prod && modif in test server
 
   const user = userBan.user;
-  const logChannel = await fetchLogChannel(userBan); //get logChannel
+  const logChannel = await fetchLogChannel(userBan.guild); //get logChannel
   const color = Colors.DarkNavy;
   const embed = setupEmbed(color, perso, user, "user"); //setup embed
   embed.addFields({ name: perso.id, value: user.id, inline: true });
@@ -866,8 +867,8 @@ export const onGuildMemberUpdate = async (_oldMember, newMember) => {
   const timeout = personality.timeout;
   const auditLog = personality.auditLog;
 
-  if (process.env.DEBUG === "no" && isTestServer(newMember)) return; //if in prod && modif in test server
-  const logChannel = await fetchLogChannel(newMember); //get logChannel
+  if (isProduction && isTestServer(newMember)) return; //if in prod && modif in test server
+  const logChannel = await fetchLogChannel(newMember.guild); //get logChannel
   const color = Colors.Orange;
   const embed = setupEmbed(color, timeout, user, "tag"); //setup embed
   const timeoutLog = await fetchAuditLog(
@@ -916,7 +917,7 @@ export const onGuildMemberRemove = async (memberKick) => {
   const personality = PERSONALITY.getAdmin(); //get personality
   const auditLog = personality.auditLog;
 
-  if (process.env.DEBUG === "no" && isTestServer(memberKick)) return; //if in prod && modif in test server
+  if (isProduction && isTestServer(memberKick)) return; //if in prod && modif in test server
   const kickLog = await fetchAuditLog(
     memberKick.guild,
     AuditLogEvent.MemberKick,
@@ -957,7 +958,7 @@ export const onGuildMemberRemove = async (memberKick) => {
         value: textRoles,
         inline: true,
       }); //add user roles if any
-    const logChannel = await fetchLogChannel(memberKick, "inAndOut"); //get logChannel
+    const logChannel = await fetchLogChannel(memberKick.guild, "inAndOut"); //get logChannel
     const messageList = await endCasesEmbed(
       userKick,
       kickLog,
@@ -983,7 +984,7 @@ export const onGuildMemberRemove = async (memberKick) => {
   ); //add user id
   if (textRoles)
     embed.addFields({ name: guildKick.roles, value: textRoles, inline: true }); //add user roles if any
-  const logChannel = await fetchLogChannel(memberKick); //get logChannel
+  const logChannel = await fetchLogChannel(memberKick.guild); //get logChannel
 
   endCasesEmbed(
     userKick,
@@ -1001,7 +1002,7 @@ export const onGuildMemberRemove = async (memberKick) => {
 export const onGuildMemberAdd = async (guildMember) => {
   const currentServer = COMMONS.fetchFromGuildId(guildMember.guild.id);
 
-  if (currentServer.name === "prod" && process.env.DEBUG === "no") {
+  if (currentServer.name === "prod" && isProduction) {
     console.log("onGuildMemberAdd", guildMember.displayName);
 
     const db = guildMember.client.db;
@@ -1055,7 +1056,7 @@ export const checkPinStatus = async (message) => {
       });
 
       //get logChannel
-      const logChannel = await fetchLogChannel(message, "thread");
+      const logChannel = await fetchLogChannel(message.guild, "thread");
       const logMessage = await channelSend(logChannel, { embeds: [embed] });
       addAdminLogs(message.client.db, logMessage.id, "frequent", 6);
     }