+++ src/admin/utils.js
@@ -20,6 +20,7 @@ import {
   checkEmbedContent,
   fetchLogChannel,
   getAdminLogs,
+  isProduction,
   parseUnixTimestamp,
   removeAdminLogs,
   removeEmote,
@@ -76,10 +77,7 @@ export const finishEmbed = async (
   stickers,
 ) => {
   const currentServer = COMMONS.getTest(); //get test data
-  if (
-    process.env.DEBUG === "no" &&
-    logChannel.guildId === currentServer.guildId
-  ) {
+  if (isProduction && logChannel.guildId === currentServer.guildId) {
     //Sil'Afian detects test in test server => return
     console.log("Sil'Afian log in Test server", eventPerso.title);
     return null;
@@ -227,9 +225,9 @@ export const processGeneralEmbed = async (
   const perso = personality[persoType];
   const aLog = personality.auditLog;
 
-  if (process.env.DEBUG === "no" && isTestServer(obj)) return; //if in prod && modif in test server
+  if (isProduction && isTestServer(obj)) return; //if in prod && modif in test server
 
-  const channel = await fetchLogChannel(obj); //get logChannel
+  const channel = await fetchLogChannel(obj.guild); //get logChannel
   const objToSend = objType === "user" ? obj.user : obj; //handle user objects case
   const embed = setupEmbed(color, perso, objToSend, embedType); //setup embed
   embed.addFields({ name: perso.id, value: objToSend.id, inline: true });
@@ -598,8 +596,7 @@ export const fetchMessageItself = async (message) => {
 const logsRemover = async (client) => {
   console.log("logsRemover");
   const db = client.db;
-  const server =
-    process.env.DEBUG === "yes" ? COMMONS.getTest() : COMMONS.getProd();
+  const server = isProduction ? COMMONS.getProd() : COMMONS.getTest();
 
   // frequent logs remove
   let type = "frequent"; //differentiate process for "frequent" and "userAD" logs
@@ -672,7 +669,7 @@ export const octagonalLog = async (object, user) => {
   if (message.partial) await fetchMessageItself(message);
 
   //basic operations
-  const logChannel = await fetchLogChannel(message); //get logChannelId
+  const logChannel = await fetchLogChannel(message.guild); //get logChannelId
   const embed = setupEmbed(
     Colors.LuminousVividPink,
     octaPerso,