+++ changelog.md
@@ -1,9 +1,12 @@
+# 1.3.0
+- Feat: add a new `spam` channel where to report bot status (online and process errors) ([#41](https://github.com/Eccleria/Sil-Afian/pull/41)) 
+
 # 1.2.0
 - CI: bump `node` requirement to 20.19.0 ([#54](https://github.com/Eccleria/Sil-Afian/pull/54))
 - Feat: add `ghostReport` `context` and `slash` commands, for silent reports to mods. Note that `octagonal` logs are still enforced ([#24](https://github.com/Eccleria/Sil-Afian/pull/24))
 - Package: Bump to latest, fixing some dependancies warnings (jest) ([#53](https://github.com/Eccleria/Sil-Afian/pull/53))
-- Refacto: move `classes` files into separate folder ([#50](https://github.com/Eccleria/Sil-Afian/pull/50)) ([#51](https://github.com/Eccleria/Sil-Afian/pull/51))
-- Update: listeners now use djs `Events` enum ([#49](https://github.com/Eccleria/Sil-Afian/pull/49)) ([#52](https://github.com/Eccleria/Sil-Afian/pull/52))
+- Refacto: move `classes` files into separate folder ([#50](https://github.com/Eccleria/Sil-Afian/issues/50)) ([#51](https://github.com/Eccleria/Sil-Afian/pull/51))
+- Update: listeners now use djs `Events` enum ([#49](https://github.com/Eccleria/Sil-Afian/issues/49)) ([#52](https://github.com/Eccleria/Sil-Afian/pull/52))
 
 # 1.1.0
 - Feat: add multiple utils used as common API wrappers ([#42](https://github.com/Eccleria/Sil-Afian/pull/42)) ([#46](https://github.com/Eccleria/Sil-Afian/pull/46))